VibesAdd/requirements.txt
python-telegram-bot==20.7
supabase==2.3.0 